[metasrv] refactor: move Cmd and LogEntry to metasrv/raft/types
