+++ metasrv/src/meta_service/mod.rs
@@ -12,10 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use cmd::Cmd;
 pub use errors::RetryableError;
 pub use errors::ShutdownError;
-pub use log_entry::LogEntry;
 pub use meta_service_impl::MetaServiceImpl;
 pub use network::Network;
 pub use raftmeta::MetaNode;
@@ -27,10 +25,10 @@ pub use crate::protobuf::meta_service_server::MetaServiceServer;
 pub use crate::protobuf::GetReply;
 pub use crate::protobuf::GetReq;
 pub use crate::protobuf::RaftMes;
+pub use crate::raft::types::cmd::Cmd;
+pub use crate::raft::types::log_entry::LogEntry;
 
-pub mod cmd;
 pub mod errors;
-pub mod log_entry;
 pub mod meta_service_impl;
 pub mod network;
 pub mod raftmeta;
